Splits tests between separate module test files
