+++ tests/test_database.py
@@ -1,0 +1,17 @@
+import pytest
+from sqlalchemy import text
+
+from database import async_session
+from main import *
+
+
+class TestDatabase:
+    @pytest.mark.asyncio
+    async def test_database_async_connection(self) -> None:
+        async with async_session() as session:
+            result = await session.execute(text("SELECT 1"))
+            value = result.scalar()
+
+            assert value == 1, (
+                "Database connection failed or returned unexpected result"
+            )
